opensource/binshot/requirements.txt
astor==0.8.1
asttokens==3.0.0
attrdict==2.0.1
backcall==0.2.0
binaryai==0.2.8
bokeh==2.3.3
cached-property==1.5.2
capstone==5.0.6
charset-normalizer==2.0.3
conf==0.4.1
cycler==0.12.1
dataclasses==0.6
decorator==5.2.1
distlib==0.3.4
executing==2.2.0
filelock==3.7.0
fsspec==2025.3.0
gast==0.3.3
h5py==2.10.0
importlib_resources==6.4.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
Keras-Applications==1.0.8
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.3.4
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
pandas==2.0.3
parso==0.8.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyelftools==0.32
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==1.3.2
scikit-learn-extra==0.2.0
scipy==1.10.1
seaborn==0.11.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
virtualenv==20.14.1
wcwidth==0.2.13
Werkzeug==2.0.1
zipp==3.20.2
